requirements.txt
crewAI
pydantic
tavily-python
scrapegraph_py
agentops
streamlit
python-dotenv
ollama